Flask_SSO/requirements.txt
pathlib
flask
google-auth
google-auth-oauthlib
vendor
Flask-Authlib-Client